requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-filter==23.3
django-redis==5.4.0
django-cors-headers==4.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-cleanup==8.0.0
django-widget-tweaks==1.5.0
tmdbsimple==2.9.1
geoip2==4.7.0
django-countries==7.5.1
django-recaptcha==4.0.0
django-cookie-consent==0.3.1
django-compressor==4.4
django-imagekit==5.0.0
django-ckeditor==6.7.0
django-mptt==0.15.0
celery==5.3.4
Unidecode==1.3.7
